issue-13: Move obstacle into it's own file
