+++ python/game/entities/__init__.py
@@ -1,7 +1,8 @@
 """"
 Game entities module
 """
+from .obstacle import Obstacle
 from .player import Player
-from .entities import Obstacle, GameState
+from .entities import GameState
 
 all__ = ['Player', 'Obstacle', 'GameState']
